refactor(wan): share family pipeline components
