+++ unirl/models/wan22/diffusion.py
@@ -37,7 +37,8 @@ import torch
 
 from unirl.models.types.diffusion import DiffusionStage, DiffusionStep
 from unirl.models.types.replay_result import ReplayResult
-from unirl.models.wan21.conditions import WAN21Conditions
+from unirl.models.wan.conditions import WANConditions
+from unirl.models.wan.geometry import wan_latent_shape
 from unirl.sde.kernels import StepStrategy
 from unirl.types.sampling import DiffusionSamplingParams, compute_trajectory_positions
 from unirl.types.segments.latent import LatentSegment, make_video_segment
@@ -48,7 +49,7 @@ from .bundle import WAN22Bundle
 _WAN_TIMESTEP_SCALE: float = 1000.0
 
 
-class WAN22DiffusionStep(DiffusionStep[WAN22Bundle, WAN21Conditions]):
+class WAN22DiffusionStep(DiffusionStep[WAN22Bundle, WANConditions]):
     """Per-step WAN 2.2 denoising kernel — stateless, dual-transformer routing.
 
     For each call, decides whether to route through the high- or
@@ -102,7 +103,7 @@ class WAN22DiffusionStep(DiffusionStep[WAN22Bundle, WAN21Conditions]):
         model: WAN22Bundle,
         sample: torch.Tensor,
         sigma: torch.Tensor,
-        conditions: WAN21Conditions,
+        conditions: WANConditions,
         *,
         guidance_scale: float,
         guidance_scale_2: Optional[float] = None,
@@ -225,7 +226,7 @@ class WAN22DiffusionStep(DiffusionStep[WAN22Bundle, WAN21Conditions]):
     def step(
         self,
         model: WAN22Bundle,
-        conditions: WAN21Conditions,
+        conditions: WANConditions,
         *,
         strategy: StepStrategy,
         sample: torch.Tensor,
@@ -269,7 +270,7 @@ class WAN22DiffusionStep(DiffusionStep[WAN22Bundle, WAN21Conditions]):
     def step_with_logp(
         self,
         model: WAN22Bundle,
-        conditions: WAN21Conditions,
+        conditions: WANConditions,
         *,
         strategy: StepStrategy,
         sample: torch.Tensor,
@@ -299,7 +300,7 @@ class WAN22DiffusionStep(DiffusionStep[WAN22Bundle, WAN21Conditions]):
         )
 
 
-class WAN22DiffusionStage(DiffusionStage[WAN21Conditions]):
+class WAN22DiffusionStage(DiffusionStage[WANConditions]):
     """WAN 2.2 T2V rollout-level diffusion stage with dual-transformer routing.
 
     Owns the SDE ``strategy`` (stateful strategies require a stable
@@ -346,16 +347,14 @@ class WAN22DiffusionStage(DiffusionStage[WAN21Conditions]):
     # ------------------------------------------------------------------
 
     def _latent_shape(self, *, num_frames: int, height: int, width: int) -> Tuple[int, int, int, int]:
-        if (int(num_frames) - 1) % self._TEMPORAL_DOWNSAMPLE != 0:
-            raise ValueError(
-                f"WAN VAE temporal_downsample={self._TEMPORAL_DOWNSAMPLE} requires "
-                f"(num_frames - 1) % {self._TEMPORAL_DOWNSAMPLE} == 0, got num_frames={num_frames}; "
-                f"valid choices: 1, 5, 9, 13, 17, 21, ..."
-            )
-        latent_t = (int(num_frames) - 1) // self.temporal_scale_factor + 1
-        latent_h = int(height) // self.vae_scale_factor
-        latent_w = int(width) // self.vae_scale_factor
-        return (self.latent_channels, latent_t, latent_h, latent_w)
+        return wan_latent_shape(
+            num_frames=num_frames,
+            height=height,
+            width=width,
+            latent_channels=self.latent_channels,
+            spatial_downsample=self.vae_scale_factor,
+            temporal_downsample=self.temporal_scale_factor,
+        )
 
     # ------------------------------------------------------------------
     # Sampling
@@ -363,7 +362,7 @@ class WAN22DiffusionStage(DiffusionStage[WAN21Conditions]):
 
     def diffuse(
         self,
-        conditions: WAN21Conditions,
+        conditions: WANConditions,
         *,
         schedule: torch.Tensor,
         params: DiffusionSamplingParams,
@@ -495,7 +494,7 @@ class WAN22DiffusionStage(DiffusionStage[WAN21Conditions]):
 
     def replay(
         self,
-        conditions: WAN21Conditions,
+        conditions: WANConditions,
         *,
         segment: LatentSegment,
         params: DiffusionSamplingParams,
@@ -574,7 +573,7 @@ class WAN22DiffusionStage(DiffusionStage[WAN21Conditions]):
 
     def predict_noise_at_step(
         self,
-        conditions: WAN21Conditions,
+        conditions: WANConditions,
         *,
         sample: torch.Tensor,
         sigma: torch.Tensor,
